+++ src/platform/api/api-router-tenant-related.js
@@ -285,6 +285,11 @@ router.post(
   AuthenticationController.isSignedIn,
   GroupBookingController.createGroupBookingReceipt,
 );
+router.post(
+  "/group-bookings/:id/invoice",
+  AuthenticationController.isSignedIn,
+  GroupBookingController.createGroupBookingInvoice,
+);
 
 // CHECKOUT
 // ========